requirements.txt
python-dotenv
marshmallow-sqlalchemy
flask-marshmallow
pybase64
requests==2.24.0
alembic==1.0.11
aniso8601==7.0.0
appnope==0.1.0
backcall==0.1.0
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
decorator==4.4.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
ipdb==0.12
ipython==7.6.1
itsdangerous==1.1.0
jedi==0.14.1
Jinja2==2.10.1
Mako==1.0.13
MarkupSafe==1.1.1
marshmallow==2.19.5
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2-binary==2.8.3
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.5
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.15.4
Flask-HTTPBasicAuth
Flask-Fixtures
